backend/llm_service/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv>=0.21.0
google-generativeai>=0.3.0
pydantic==2.0.2
httpx==0.24.1
pydantic-settings==2.0.2
PyYAML>=6.0

